mpconfigport.h: Enable running pre-compiled scripts from the file system
This feature is supported in the code, but not enabled yet. Enabling it
allows to run code pre-compiled with mpy-cross without the need to embed it
into the formware as frozen bytecode. Running bytecode from the file system
saves the compiling phase, allowing larger scripts and giving faster load time.