Lora: Changes for the EU433 support
First commit for EU433
These changes allow the use of the region EU433. It is not fully tested yet,
but at least ABO Join, uplink and downlink in RX2 was confirmed to work.
Some modes could not be tested due to limitations at the LoRa Servers,
and maybe the timing is still wrong.

This commit is based on Version 1.9.0.b3 with the PR allowing to disable
Ctrl-F Reset in a script.

Changed files:
application.mk  (just added the flag -DEU433)
mods/modlora.c  (handling all of the EU433 options)
../lib/lora/mac/region/Region.c  (just for manual channel add with frequency)
../lib/lora/mac/region/RegionEU433.c (just for manual channel add with frequency)
../lib/lora/mac/region/RegionEU433.h (just for manual channel add with frequency)