BLE characteristic update messages are lost if they sent too frequently
In modbt.c it can happen that received packets are lost, because the user registered callback is not called fast enough (via gatts_char_callback_handler()), which means it can happen that the value of a characteristic is updated 2 times in a row (via gatts_event_handler()) without giving the possibility to the callback to run and fetch the new value. In this specific example the first callback will show the 2nd change, and the 2nd callback will show that actually nothing happened. The 1st change gets lost.
The fix of this issue breaks backward compatibility, because the prototype of the user registered Python callback is changed:
OLD: callback(characteristic)
NEW: callback(characteristic, (event, value))  → (event, value) is a tuple, where data is bytearray containing the value of the request if the “event” is WRITE, otherwise “value” is “None”. “event” shows the correct event any time.