MicroPython release 1.20.2.rc2 based on Espressif IDF 3.3.1 (#394)
* BLE: Allow static passwords, remove bonded when pin changes

* mods/pybadc.c: Fix the argument handling of bits=x for adc.init()

Before, the only accepted argument & value for adc.init() was bits=12, and even that was
discarded. With that change, the function meets the documentation and works as expected.

* mods/pyblte.c: Fix the lenght field for AT commands

At three places the length field was not set when AT commands werde
issued, causing lte.attach() with the band=xx option to fail.
Also, affected: lte_get_modem_version()

* Support Sigfox registration for Pybytes

* PYFW-390: Update .gitignore to allow esp32/lib libraries to be updated

* MDNS advertisement works

* Adding text to advertisement works

* Query works

* Free up internal query results

* Change host_name to hostname

* Rename text to txt

* Adding libmdns.a to esp32/lib

* Adding scripts for PyJTAG and short Readme

* update sigfox libraries

* Replace xQueueSendFromISR call with xQueueSend when called from BLE event

Use the callback queue to notify the LoRa Timer Thread
Do context switch in TimerCallback only if needed

* Replacing "switch-case" with "if-else if" in Region.c to avoid generated Assembly instructions which are reading the Flash from IRAM functions causing bad00bad issue

* BLE characteristic update messages are lost if they sent too frequently

In modbt.c it can happen that received packets are lost, because the user registered callback is not called fast enough (via gatts_char_callback_handler()), which means it can happen that the value of a characteristic is updated 2 times in a row (via gatts_event_handler()) without giving the possibility to the callback to run and fetch the new value. In this specific example the first callback will show the 2nd change, and the 2nd callback will show that actually nothing happened. The 1st change gets lost.
The fix of this issue breaks backward compatibility, because the prototype of the user registered Python callback is changed:
OLD: callback(characteristic)
NEW: callback(characteristic, (event, value))  → (event, value) is a tuple, where data is bytearray containing the value of the request if the “event” is WRITE, otherwise “value” is “None”. “event” shows the correct event any time.

* PYFW-394: mod_ssl_setup_socket() allocates memory while GIL is not locked

* PYFW-401: Mutex of LFS can be locked in wlan_read_file or mod_ssl_read_file

* PYFW-402: wlan_do_connect() allocates memory while GIL is not locked

PYFW-403: mod_network_register_nic() using MicroPython functions outside of GIL in wlan_setup_ap()

* PYFW-404: bt_connect_helper() uses MicroPython functions outside of GIL

* PYFW-405: lte_send_at_cmd() uses MicroPython functions outside of GIL

* refactor lte_add_band() out of lte_attach()

* add bands=() parameter to lte.attach()

this way the user can select a set of bands they want to use when attaching

* PYFW-391: Add changes needed for esp-idf v3.3.1

* Updating README.md

* Update sdkconfig.h and libraries when secure boot is ON, but CONFIG_SECURE_BOOT_ENABLED is removed from sdkconfig.h

* Add checker which forbids compilation when CONFIG_SECURE_BOOT_ENABLED is defined but SECURE=on is not

* Add pybytes_on_boot functionality (#87)

* Update pycom_version.h

Update version to 1.20.2.rc2 for public release

* Update Pybytes to version 1.3.1

Co-authored-by: oligauc <oliver@pycom.io>
Co-authored-by: Islam Wahdan <islam@pycom.io>
Co-authored-by: robert-hh <robert@hammelrath.com>
Co-authored-by: Géza Husi <husigeza91@gmail.com>
Co-authored-by: Jirka Krepl <jiri.krepl@gmail.com>
Co-authored-by: doniks <p.putz@yahoo.de>