lora: Fix losing of multicast params
we're allocating MulticastParams_t inside garbage collected memory, then pass it into the C-level implementation. It seems the garbage collector cannot keep track of the usage this way, so eventually recycles the memory. At that point we lose the multicast addresses and when we check it in

LoraMac.c: 842
        case FRAME_TYPE_DATA_CONFIRMED_DOWN:
        case FRAME_TYPE_DATA_UNCONFIRMED_DOWN:

and further we end up dereferencing those pointers still and then it's a bit random whether we core dump due to illegal memory address, or simply ignore the lora packets and become "deaf" to them